chore: better setup for `eslint-plugin-crap`
Make use of NPM workspaces instead of file linking, because
it handles dependencies better. Before, dependencies of `eslint-plugin-crap`
kept moving between the plugin and root directory with every install.

Also removes source map and type definitions for the plugin, as those
are not going to be imported directly and cause problems with portability (TS2742).